requirements.txt
openai
chromadb
flask
flask-cors
pyyaml